+++ packages/play/src/index.js
@@ -39,4 +39,4 @@ export {
 } from './sessions.js'
 export { resolvePlayPath, splitRelativeSegments } from './paths.js'
 export { createOperationContext, operationLogConstants } from './operation-log.js'
-export { detachSessionFromTimeline, PlayMembershipService } from './membership.js'
+export { detachSessionFromTimeline, PlayMembershipService, sessionIdsInTimeline } from './membership.js'